+++ core/server/api/canary/membersStripeConnect.js
@@ -1,7 +1,4 @@
 const membersService = require('../../services/members');
-const config = require('../../../shared/config');
-const urlUtils = require('../../../shared/url-utils');
-const {BadRequestError} = require('@tryghost/errors');
 
 module.exports = {
     docName: 'members_stripe_connect',
@@ -18,13 +15,6 @@ module.exports = {
             }
         },
         query(frame) {
-            const siteUrl = urlUtils.getSiteUrl();
-            const productionMode = config.get('env') === 'production';
-            const siteUrlUsingSSL = /^https/.test(siteUrl);
-            const cannotConnectToStripe = productionMode && !siteUrlUsingSSL;
-            if (cannotConnectToStripe) {
-                throw new BadRequestError('Cannot connect to stripe unless site is using https://');
-            }
             // This is something you have to do if you want to use the "framework" with access to the raw req/res
             frame.response = async function (req, res) {
                 function setSessionProp(prop, val) {
